Move `gix` project source implementation to enum variant
