+++ packages/ploys/src/project/source/git/mod.rs
@@ -3,23 +3,22 @@
 //! This module contains the utilities related to local Git project management.
 
 mod error;
+mod gix;
 
-use std::io;
 use std::path::{Path, PathBuf};
 
-use gix::remote::Direction;
-use gix::traverse::tree::Recorder;
-use gix::Repository;
 use url::Url;
 
 pub use self::error::{Error, GitError};
+pub use self::gix::Gix;
 
 use super::Source;
 
 /// The local Git repository source.
 #[derive(Clone, Debug)]
-pub struct Git {
-    repository: Repository,
+pub enum Git {
+    /// The `gix` source for basic `git` operations.
+    Gix(Gix),
 }
 
 impl Git {
@@ -28,9 +27,7 @@ impl Git {
     where
         P: AsRef<Path>,
     {
-        Ok(Self {
-            repository: gix::open(path.as_ref())?,
-        })
+        Ok(Self::Gix(Gix::new(path)?))
     }
 }
 
@@ -46,77 +43,29 @@ impl Source for Git {
     }
 
     fn get_name(&self) -> Result<String, Self::Error> {
-        let path = self.repository.path().join("..").canonicalize()?;
-
-        if let Some(file_stem) = path.file_stem() {
-            return Ok(file_stem.to_string_lossy().to_string());
+        match self {
+            Self::Gix(gix) => gix.get_name(),
         }
-
-        Err(Error::Io(io::Error::new(
-            io::ErrorKind::Other,
-            "Invalid directory",
-        )))
     }
 
     fn get_url(&self) -> Result<Url, Self::Error> {
-        match self
-            .repository
-            .find_default_remote(Direction::Push)
-            .transpose()?
-        {
-            Some(remote) => match remote.url(Direction::Push) {
-                Some(url) => Ok(url
-                    .to_bstring()
-                    .to_string()
-                    .parse()
-                    .expect("A repository URL should be valid")),
-                None => Err(Error::remote_not_found()),
-            },
-            None => Err(Error::remote_not_found()),
+        match self {
+            Self::Gix(gix) => gix.get_url(),
         }
     }
 
     fn get_files(&self) -> Result<Vec<PathBuf>, Self::Error> {
-        let tree = self
-            .repository
-            .rev_parse_single("HEAD")?
-            .object()?
-            .peel_to_tree()?;
-
-        let mut recorder = Recorder::default();
-
-        tree.traverse().breadthfirst(&mut recorder)?;
-
-        let mut entries = recorder
-            .records
-            .into_iter()
-            .filter(|entry| entry.mode.is_blob())
-            .map(|entry| PathBuf::from(entry.filepath.to_string()))
-            .collect::<Vec<_>>();
-
-        entries.sort();
-
-        Ok(entries)
+        match self {
+            Self::Gix(gix) => gix.get_files(),
+        }
     }
 
     fn get_file_contents<P>(&self, path: P) -> Result<Vec<u8>, Self::Error>
     where
         P: AsRef<Path>,
     {
-        let mut tree = self
-            .repository
-            .rev_parse_single("HEAD")?
-            .object()?
-            .peel_to_tree()?;
-
-        let entry = tree
-            .peel_to_entry_by_path(path)?
-            .ok_or_else(|| io::Error::from(io::ErrorKind::NotFound))?;
-
-        if entry.mode().is_blob() {
-            Ok(entry.object()?.detached().data)
-        } else {
-            Err(io::Error::from(io::ErrorKind::NotFound))?
+        match self {
+            Self::Gix(gix) => gix.get_file_contents(path),
         }
     }
 }
